requirements.txt
fastapi
uvicorn
pydantic
joblib
scikit-learn==1.6.1
pandas
numpy
requests
python-dotenv
streamlit